requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
PyYAML==6.0.2
Werkzeug==3.1.2
pytest==8.3.4